Restore func_002ed430 with a consistent u16 persona-id contract
The parked body needed a block-scope func_00310a10(u8 *, u16) declaration
that conflicted with its s32 definition. The real contract is u16 all the
way down: func_00310a10, func_00109280, func_00109220 and func_003026c0
take u16, and func_001092f0 returns u16. MWCC then passes the lhu class id
raw to both callees, as retail does.

The g_data accessors keep retail's two separate zero-extensions through a
(u32) row offset. Caller declarations in ten units now agree with the
definitions. Two shdPersona call sites and one code1_0036 local were
retyped, with unchanged bytes. Every affected unit keeps or improves its
status, and y_fclCombine goes to 33 MATCH / 8 ASM.

+++ src/Event/Fcl/y_fclCombineDraw.c
@@ -139,7 +139,7 @@ extern s32 func_00279350(f32, f32, f32, s32, s32, s32, s32, s32, s32, u8 *);
 extern int func_00275820(f32, f32, f32, int, s8, int, const char *, int, int, void *, int);
 extern void func_0034a640(s32, u16, s32);
 extern s32 func_0034a630(u8 *);
-extern u8 func_00109280(s32);
+extern u8 func_00109280(u16);
 extern void func_0011d1d0(u8 *, f32);
 extern s8 func_00331640(void);
 extern void func_00330e50(s32, FclVec2, f32, s32, u8, s32, f32, f32, void *);
